Implement PID controller for steering

+++ sw/peripherals-board/src/managers/vec3f.hpp
@@ -2,8 +2,8 @@
 
 struct Vec3f{
 
-  float x = 0;
-  float y = 0;
-  float z = 0;
+  double x = 0;
+  double y = 0;
+  double z = 0;
 
 };
